refactor(app): move TextEditor out of app state into ainb-app
TextEditor is a self-contained cursor-aware text buffer used by the boss
mode prompt and the new-session Configure form. It moves to
ainb_app::text_editor so the Configure state can follow it;
app::state re-exports it under its old path.
