+++ ainb-tui/crates/ainb-app/src/lib.rs
@@ -30,6 +30,7 @@ pub mod providers;
 pub mod rtk;
 pub mod self_exec_guard;
 pub mod setup;
+pub mod text_editor;
 pub mod tmux;
 pub mod usage_cache;
 pub mod viewport;
